Extract CenterContainer into its own component
