+++ application/src/components/Containers.js
@@ -1,5 +1,6 @@
 // React
 import PropTypes from 'prop-types'
+import { Container } from 'react-bootstrap';
 
 // These FRow and FCol components are just container divs with a specific class attached to them.
 // The "complicated stuff" is just so that it preserves props and classes given, and allow children within them.
@@ -28,8 +29,21 @@ function FCol(props) {
 }
 
 FCol.propTypes = {
-  children: PropTypes.node, // <- this is the HTML elements inside this component
+  children: PropTypes.node,
   className: PropTypes.string
 }
 
-export {FRow, FCol};
+function CenterContainer(props) {
+  return (
+    <div className="d-flex">
+      <Container className="outer-container w-auto">
+        {props.children}
+      </Container>
+    </div>
+  );
+}
+CenterContainer.propTypes = {
+  children: PropTypes.node
+}
+
+export {FRow, FCol, CenterContainer};
